backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
langchain==0.0.335
langchain-openai==0.0.2
faiss-cpu==1.13.2
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sentence-transformers==2.2.2
tiktoken==0.5.1
aiofiles==23.2.1
